// src/app/models/Vehicle.ts
export class Vehicle {
    id?: string;
    name?: string;
    serialNumber?: string;
    licensePlate?: string;
    creationYear?: number;
    oem?: string;
    model?: string;

    // constructor(idNumber: string, name: string, serialNumber: string, licensePlate: string, creationYear: number, oem: string, model: string) {
    //     this.idNumber = idNumber;
    //     this.name = name;
    //     this.serialNumber = serialNumber;
    //     this.licensePlate = licensePlate;
    //     this.creationYear = creationYear;
    //     this.oem = oem;
    //     this.model = model;
    // }
}